src/main/java/com/oli/Services/VideoService.java
package com.oli.Services;

import com.oli.Entities.Video;
import com.oli.Repository.VideoRepository;
import lombok.extern.slf4j.Slf4j;
import org.springframework.data.repository.PagingAndSortingRepository;
import org.springframework.stereotype.Service;

/**
 * Author: Oliver
 */
@Slf4j
@Service
public class VideoService extends ServiceAbstract<Video, Long> {
    private final VideoRepository repository;

    public VideoService(VideoRepository repository) {
        this.repository = repository;
    }

    @Override
    PagingAndSortingRepository<Video, Long> getRepository() {
        return repository;
    }

    @Override
    public long countByCreatedBy(String user) {
        return repository.countByEmbeddedAudits_CreatedBy(user);
    }
}
